fix weird warning in Lake_McClure_Water_Demand.py

+++ models/merced/_parameters/Lake_McClure_Water_Demand.py
@@ -22,7 +22,7 @@ class Lake_McClure_Water_Demand(WaterLPParameter):
         curves_af = self.model.tables["Lake McClure/Guide Curve"]
         max_volume_mcm = self.model.nodes[self.res_name].max_volume.get_value(scenario_index)
         date_str = '1900-{:02}-{:02}'.format(timestep.month, timestep.day)
-        target_mcm = curves_af.at[date_str, self.wyt] * 1233.5 / 1e6
+        target_mcm = float(curves_af.at[date_str, self.wyt] * 1233.5 / 1e6)
         target_fraction = min(target_mcm / max_volume_mcm, 1.0)
         return target_fraction
